make reboot work
