+++ pico-rv32ima/main.c
@@ -11,8 +11,6 @@
 #include "console.h"
 #include "terminal.h"
 
-const char *const imageFilename = "0:Image";
-
 void core1_entry();
 
 int main()
@@ -46,13 +44,8 @@ void core1_entry()
 	if (FR_OK != fr)
 		console_panic("SD mount error: %s (%d)\n\r", FRESULT_str(fr), fr);
 
-	fr = loadFileIntoRAM(imageFilename, 0);
-	if (FR_OK != fr)
-		console_panic("Error loading image: %s (%d)\n", FRESULT_str(fr), fr);
-	console_printf("Image loaded sucessfuly!\n\r");
+	int c = rvEmulator();
 
-	rvEmulator();
-
-	while (true)
-		tight_loop_contents();
+	while (c == EMU_REBOOT)
+		c = rvEmulator();
 }
